pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hltrader"
version = "0.1.0"
description = "Stop-loss and risk management CLI for Hyperliquid"
requires-python = ">=3.10"
dependencies = [
    "hyperliquid-python-sdk>=0.9.0",
    "typer[all]>=0.9.0",
    "rich>=13.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "eth-account>=0.9.0",
    "requests>=2.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0"]

[project.scripts]
hl = "hltrader.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["hltrader*"]
